requirements.txt
streamlit>=1.35.0
openai>=1.30.0
yfinance>=0.2.40
pandas>=2.0.0
plotly>=5.20.0
sec-api>=1.0.0
python-dateutil>=2.9.0